feat: moving the actions and actions registry to alpha services Signed-off-by: benjdlambert <ben@blam.sh>
