+++ packages/backend-defaults/package.json
@@ -20,8 +20,6 @@
   "license": "Apache-2.0",
   "exports": {
     ".": "./src/index.ts",
-    "./actions": "./src/entrypoints/actions/index.ts",
-    "./actionsRegistry": "./src/entrypoints/actionsRegistry/index.ts",
     "./auditor": "./src/entrypoints/auditor/index.ts",
     "./auth": "./src/entrypoints/auth/index.ts",
     "./cache": "./src/entrypoints/cache/index.ts",
@@ -41,18 +39,13 @@
     "./scheduler": "./src/entrypoints/scheduler/index.ts",
     "./urlReader": "./src/entrypoints/urlReader/index.ts",
     "./userInfo": "./src/entrypoints/userInfo/index.ts",
+    "./alpha": "./src/alpha/index.ts",
     "./package.json": "./package.json"
   },
   "main": "src/index.ts",
   "types": "src/index.ts",
   "typesVersions": {
     "*": {
-      "actions": [
-        "src/entrypoints/actions/index.ts"
-      ],
-      "actionsRegistry": [
-        "src/entrypoints/actionsRegistry/index.ts"
-      ],
       "auditor": [
         "src/entrypoints/auditor/index.ts"
       ],
@@ -109,6 +102,9 @@
       ],
       "userInfo": [
         "src/entrypoints/userInfo/index.ts"
+      ],
+      "alpha": [
+        "src/alpha/index.ts"
       ],
       "package.json": [
         "package.json"
