--- packages/backend-defaults/report-actionsRegistry.api.md
@@ -1,17 +1,0 @@
-## API Report File for "@backstage/backend-defaults"
-
-> Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
-
-```ts
-import { ActionsRegistryService } from '@backstage/backend-plugin-api';
-import { ServiceFactory } from '@backstage/backend-plugin-api';
-
-// @public (undocumented)
-export const actionsRegistryServiceFactory: ServiceFactory<
-  ActionsRegistryService,
-  'plugin',
-  'singleton'
->;
-
-// (No @packageDocumentation comment for this package)
-```
