+++ packages/backend-defaults/report-alpha.api.md
@@ -3,8 +3,16 @@
 > Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
 
 ```ts
-import { ActionsService } from '@backstage/backend-plugin-api';
+import { ActionsRegistryService } from '@backstage/backend-plugin-api/alpha';
+import { ActionsService } from '@backstage/backend-plugin-api/alpha';
 import { ServiceFactory } from '@backstage/backend-plugin-api';
+
+// @public (undocumented)
+export const actionsRegistryServiceFactory: ServiceFactory<
+  ActionsRegistryService,
+  'plugin',
+  'singleton'
+>;
 
 // @public (undocumented)
 export const actionsServiceFactory: ServiceFactory<
