+++ packages/backend-defaults/src/CreateBackend.ts
@@ -35,12 +35,12 @@ import { schedulerServiceFactory } from '@backstage/backend-defaults/scheduler';
 import { urlReaderServiceFactory } from '@backstage/backend-defaults/urlReader';
 import { userInfoServiceFactory } from '@backstage/backend-defaults/userInfo';
 import { eventsServiceFactory } from '@backstage/plugin-events-node';
-import { actionsRegistryServiceFactory } from './entrypoints/actionsRegistry';
-import { actionsServiceFactory } from './entrypoints/actions';
-
-export const defaultServiceFactories = [
+import {
   actionsRegistryServiceFactory,
   actionsServiceFactory,
+} from '@backstage/backend-defaults/alpha';
+
+export const defaultServiceFactories = [
   auditorServiceFactory,
   authServiceFactory,
   cacheServiceFactory,
@@ -61,6 +61,10 @@ export const defaultServiceFactories = [
   userInfoServiceFactory,
   urlReaderServiceFactory,
   eventsServiceFactory,
+
+  // alpha services
+  actionsRegistryServiceFactory,
+  actionsServiceFactory,
 ];
 
 /**
